fixed folder naming for Android tutorial 3
