+++ android/tutorials/tutorial3-completed/README.md
@@ -1,10 +1,12 @@
 # RIB Tutorial 3: RIB DI and communication
 
-This project is the completed code for [tutorial 3](https://github.com/uber/RIBs/wiki/Android-Tutorial-3).
+This project is the completed code for [tutorial 2](https://github.com/uber/RIBs/wiki/Android-Tutorial-2) and the starting point for [tutorial 3](https://github.com/uber/RIBs/wiki/Android-Tutorial-3).
 
 ### Getting started
 Run the code by
 
 ```
-./gradlew :tutorials:tutorial3-completed:installDebug
+./gradlew :tutorials:tutorial3:installDebug
 ```
+
+Then follow the steps described in [tutorial 3](https://github.com/uber/RIBs/wiki/Android-Tutorial-3) on the RIBs wiki.
