+++ android/tutorials/tutorial3-completed/build.gradle
@@ -47,8 +47,8 @@ dependencies {
     compile project(":libraries:rib-android")
     compile deps.external.dagger
     compile deps.external.rxbinding
-    compile 'com.google.guava:guava:11.0.2'
     compile 'com.android.support:percent:25.0.1'
+    compile 'com.google.guava:guava:11.0.2'
     provided deps.apt.javax
     testCompile project(":libraries:rib-test-utils")
 }
