+++ android/tutorials/tutorial3-completed/src/main/java/com/uber/rib/root/logged_in/LoggedInInteractor.java
@@ -22,12 +22,16 @@ import com.uber.rib.core.EmptyPresenter;
 import com.uber.rib.core.Interactor;
 import com.uber.rib.core.RibInteractor;
 import com.uber.rib.root.logged_in.off_game.OffGameInteractor;
+import com.uber.rib.root.logged_in.tic_tac_toe.TicTacToeInteractor;
+import javax.inject.Inject;
 
 /**
  * Coordinates Business Logic for {@link LoggedInScope}.
  */
 @RibInteractor
 public class LoggedInInteractor extends Interactor<EmptyPresenter, LoggedInRouter> {
+
+  @Inject @LoggedInBuilder.LoggedInInternal MutableScoreStream scoreStream;
 
   @Override
   protected void didBecomeActive(@Nullable Bundle savedInstanceState) {
@@ -38,12 +42,32 @@ public class LoggedInInteractor extends Interactor<EmptyPresenter, LoggedInRoute
 
   }
 
+  @Override
+  protected void willResignActive() {
+    super.willResignActive();
+
+    // TODO: Perform any required clean up here, or delete this method entirely if not needed.
+  }
+
   class OffGameListener implements OffGameInteractor.Listener {
 
     @Override
     public void onStartGame() {
       getRouter().detachOffGame();
       getRouter().attachTicTacToe();
+    }
+  }
+
+  class TicTacToeListener implements TicTacToeInteractor.Listener {
+
+    @Override
+    public void gameWon(@Nullable String winner) {
+      if (winner != null) {
+        scoreStream.addVictory(winner);
+      }
+
+      getRouter().detachTicTacToe();
+      getRouter().attachOffGame();
     }
   }
 }
