+++ android/tutorials/tutorial3-completed/src/main/java/com/uber/rib/root/logged_in/MutableScoreStream.java
