+++ android/tutorials/tutorial3-completed/src/main/java/com/uber/rib/root/logged_out/LoggedOutInteractor.java
@@ -18,7 +18,6 @@ package com.uber.rib.root.logged_out;
 
 import android.support.annotation.Nullable;
 import android.support.v4.util.Pair;
-
 import com.uber.rib.core.Bundle;
 import com.uber.rib.core.Interactor;
 import com.uber.rib.core.RibInteractor;
@@ -41,12 +40,12 @@ public class LoggedOutInteractor
   protected void didBecomeActive(@Nullable Bundle savedInstanceState) {
     super.didBecomeActive(savedInstanceState);
     presenter
-        .loginName()
+        .playerNames()
         .subscribe(new Consumer<Pair<String, String>>() {
           @Override
           public void accept(Pair<String, String> names) throws Exception {
             if (!isEmpty(names.first) && !isEmpty(names.second)) {
-              listener.login(names.first, names.second);
+              listener.requestLogin(names.first, names.second);
             }
           }
         });
@@ -61,11 +60,12 @@ public class LoggedOutInteractor
    */
   interface LoggedOutPresenter {
 
-    Observable<Pair<String, String>> loginName();
+    Observable<Pair<String, String>> playerNames();
   }
 
   public interface Listener {
-    void login(String userNameA, String userNameB);
+
+    void requestLogin(String playerOne, String playerTwo);
   }
 
 }
