+++ android/tutorials/tutorial3-completed/src/main/java/com/uber/rib/root/logged_out/LoggedOutView.java
@@ -4,7 +4,15 @@ import android.content.Context;
 import android.support.annotation.Nullable;
 import android.support.v4.util.Pair;
 import android.util.AttributeSet;
+import android.widget.Button;
+import android.widget.EditText;
 import android.widget.LinearLayout;
+import com.jakewharton.rxbinding2.view.RxView;
+import com.uber.rib.core.Initializer;
+import com.uber.rib.tutorial1.R;
+import io.reactivex.Observable;
+import io.reactivex.functions.Function;
+
 /*
  * Copyright (C) 2017. Uber Technologies
  *
@@ -21,16 +29,14 @@ import android.widget.LinearLayout;
  * limitations under the License.
  */
 
-import android.widget.TextView;
-import com.jakewharton.rxbinding2.view.RxView;
-import com.uber.rib.tutorial1.R;
-import io.reactivex.Observable;
-import io.reactivex.functions.Function;
-
 /**
  * Top level view for {@link LoggedOutBuilder.LoggedOutScope}.
  */
 public class LoggedOutView extends LinearLayout implements LoggedOutInteractor.LoggedOutPresenter {
+
+  private Button loginButton;
+  private EditText playerOneEditText;
+  private EditText playerTwoEditText;
 
   public LoggedOutView(Context context) {
     this(context, null);
@@ -44,16 +50,25 @@ public class LoggedOutView extends LinearLayout implements LoggedOutInteractor.L
     super(context, attrs, defStyle);
   }
 
+  @Initializer
   @Override
-  public Observable<Pair<String, String>> loginName() {
+  protected void onFinishInflate() {
+    super.onFinishInflate();
+    playerOneEditText = (EditText) findViewById(R.id.player_one_name);
+    playerTwoEditText = (EditText) findViewById(R.id.player_two_name);
+    loginButton = (Button) findViewById(R.id.login_button);
+  }
+
+  @Override
+  public Observable<Pair<String, String>> playerNames() {
     return RxView.clicks(findViewById(R.id.login_button))
-        .map(new Function<Object, Pair<String, String>>() {
-          @Override
-          public Pair<String, String> apply(Object o) throws Exception {
-            TextView playerNameOne = (TextView) findViewById(R.id.player_name_1);
-            TextView playerNameTwo = (TextView) findViewById(R.id.player_name_2);
-            return Pair.create(playerNameOne.getText().toString(), playerNameTwo.getText().toString());
-          }
-        });
+        .map(
+            new Function<Object, Pair<String, String>>() {
+              @Override
+              public Pair<String, String> apply(Object irrelevant) throws Exception {
+                return new Pair<>(
+                    playerOneEditText.getText().toString(), playerTwoEditText.getText().toString());
+              }
+            });
   }
 }
