+++ android/tutorials/tutorial3-completed/src/test/java/com/uber/rib/root/logged_in/LoggedInInteractorTest.java
@@ -13,6 +13,7 @@ public class LoggedInInteractorTest extends RibTestBasePlaceholder {
 
     @Mock EmptyPresenter presenter;
     @Mock LoggedInRouter router;
+    @Mock MutableScoreStream scoreStream;
 
     private LoggedInInteractor interactor;
 
@@ -20,6 +21,6 @@ public class LoggedInInteractorTest extends RibTestBasePlaceholder {
     public void setup() {
         MockitoAnnotations.initMocks(this);
 
-        interactor = TestLoggedInInteractor.create();
+        interactor = TestLoggedInInteractor.create(scoreStream);
     }
 }
