+++ android/tutorials/tutorial3-completed/src/test/java/com/uber/rib/root/logged_in/off_game/OffGameInteractorTest.java
@@ -3,6 +3,7 @@ package com.uber.rib.root.logged_in.off_game;
 import com.uber.rib.core.RibTestBasePlaceholder;
 import com.uber.rib.core.InteractorHelper;
 
+import com.uber.rib.root.logged_in.ScoreStream;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -10,9 +11,15 @@ import org.mockito.MockitoAnnotations;
 
 public class OffGameInteractorTest extends RibTestBasePlaceholder {
 
+    private final String playerOne = "playerOne";
+    private final String playerTwo = "playerTwo";
+
     @Mock OffGameInteractor.Listener listener;
     @Mock OffGameInteractor.OffGamePresenter presenter;
     @Mock OffGameRouter router;
+    @Mock ScoreStream scoreStream;
+
+
 
     private OffGameInteractor interactor;
 
@@ -20,6 +27,11 @@ public class OffGameInteractorTest extends RibTestBasePlaceholder {
     public void setup() {
         MockitoAnnotations.initMocks(this);
 
-        interactor = TestOffGameInteractor.create(listener, presenter);
+        interactor = TestOffGameInteractor.create(
+            playerOne,
+            playerTwo,
+            listener,
+            presenter,
+            scoreStream);
     }
 }
