+++ android/tutorials/tutorial3-completed/src/test/java/com/uber/rib/root/logged_out/LoggedOutInteractorTest.java
@@ -1,12 +1,11 @@
 package com.uber.rib.root.logged_out;
 
-import android.support.v4.util.Pair;
-
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import android.support.v4.util.Pair;
 import com.uber.rib.core.InteractorHelper;
 import com.uber.rib.core.RibTestBasePlaceholder;
 import io.reactivex.Observable;
@@ -30,22 +29,23 @@ public class LoggedOutInteractorTest extends RibTestBasePlaceholder {
     interactor = TestLoggedOutInteractor.create(listener, presenter);
   }
 
-
   @Test
   public void attach_whenViewEmitsName_shouldCallListener() {
-    when(presenter.loginName()).thenReturn(Observable.just(Pair.create("fakename", "fakename")));
+    String fakeName1 = "1";
+    String fakeName2 = "2";
+    when(presenter.playerNames()).thenReturn(Observable.just(new Pair<>(fakeName1, fakeName2)));
 
     InteractorHelper.attach(interactor, presenter, router, null);
 
-    verify(listener).login(any(String.class), any(String.class));
+    verify(listener).requestLogin(any(String.class), any(String.class));
   }
 
   @Test
   public void attach_whenViewEmitsEmptyName_shouldNotCallListener() {
-    when(presenter.loginName()).thenReturn(Observable.just(Pair.create("", "")));
+    when(presenter.playerNames()).thenReturn(Observable.just(new Pair<String, String>("", "")));
 
     InteractorHelper.attach(interactor, presenter, router, null);
 
-    verify(listener, never()).login(any(String.class), any(String.class));
+    verify(listener, never()).requestLogin(any(String.class), any(String.class));
   }
 }
