+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/RootInteractor.java
@@ -40,10 +40,10 @@ public class RootInteractor extends Interactor<RootInteractor.RootPresenter, Roo
   class LoggedOutListener implements LoggedOutInteractor.Listener {
 
     @Override
-    public void requestLogin(String playerOne, String playerTwo) {
+    public void login(String userNameA, String userNameB) {
       // Switch to logged in. Let’s just ignore userName for now.
       getRouter().detachLoggedOut();
-      getRouter().attachLoggedIn(playerOne, playerTwo);
+      getRouter().attachLoggedIn();
     }
   }
