+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/RootRouter.java
@@ -57,7 +57,7 @@ public class RootRouter extends ViewRouter<RootView, RootInteractor, RootBuilder
     }
   }
 
-  void attachLoggedIn(String playerOne, String playerTwo) {
-    attachChild(loggedInBuilder.build(playerOne, playerTwo));
+  void attachLoggedIn() {
+    attachChild(loggedInBuilder.build());
   }
 }
