+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/logged_in/off_game/OffGameBuilder.java
@@ -22,13 +22,11 @@ import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import com.uber.rib.core.InteractorBaseComponent;
 import com.uber.rib.core.ViewBuilder;
-import com.uber.rib.root.logged_in.ScoreStream;
 import com.uber.rib.tutorial1.R;
 import dagger.Binds;
 import dagger.BindsInstance;
 import dagger.Provides;
 import java.lang.annotation.Retention;
-import javax.inject.Named;
 import javax.inject.Qualifier;
 import javax.inject.Scope;
 
@@ -66,13 +64,7 @@ public class OffGameBuilder
 
   public interface ParentComponent {
 
-    @Named("player_one") String playerOne();
-
-    @Named("player_two") String playerTwo();
-
     OffGameInteractor.Listener listener();
-
-    ScoreStream scoreStream();
   }
 
   @dagger.Module
