+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/logged_in/off_game/OffGameInteractor.java
@@ -17,16 +17,12 @@
 package com.uber.rib.root.logged_in.off_game;
 
 import android.support.annotation.Nullable;
-import com.google.common.collect.ImmutableMap;
-import com.uber.autodispose.ObservableScoper;
 import com.uber.rib.core.Bundle;
 import com.uber.rib.core.Interactor;
 import com.uber.rib.core.RibInteractor;
-import com.uber.rib.root.logged_in.ScoreStream;
 import io.reactivex.Observable;
 import io.reactivex.functions.Consumer;
 import javax.inject.Inject;
-import javax.inject.Named;
 
 /**
  * Coordinates Business Logic for {@link OffGameScope}.
@@ -35,36 +31,19 @@ import javax.inject.Named;
 public class OffGameInteractor
     extends Interactor<OffGameInteractor.OffGamePresenter, OffGameRouter> {
 
-  @Inject @Named("player_one") String playerOne;
-  @Inject @Named("player_two") String playerTwo;
   @Inject Listener listener;
   @Inject OffGamePresenter presenter;
-  @Inject ScoreStream scoreStream;
-
 
   @Override
   protected void didBecomeActive(@Nullable Bundle savedInstanceState) {
     super.didBecomeActive(savedInstanceState);
 
-    presenter.setPlayerNames(playerOne, playerTwo);
     presenter
         .startGameRequest()
         .subscribe(new Consumer<Object>() {
           @Override
           public void accept(Object object) throws Exception {
             listener.onStartGame();
-          }
-        });
-
-    scoreStream.scores()
-        .to(new ObservableScoper<ImmutableMap<String, Integer>>(this))
-        .subscribe(new Consumer<ImmutableMap<String,Integer>>() {
-          @Override
-          public void accept(ImmutableMap<String, Integer> scores)
-              throws Exception {
-            Integer playerOneScore = scores.get(playerOne);
-            Integer playerTwoScore = scores.get(playerTwo);
-            presenter.setScores(playerOneScore, playerTwoScore);
           }
         });
   }
@@ -78,10 +57,6 @@ public class OffGameInteractor
    * Presenter interface implemented by this RIB's view.
    */
   interface OffGamePresenter {
-
-    void setPlayerNames(String playerOne, String playerTwo);
-
-    void setScores(Integer playerOneScore, Integer playerTwoScore);
 
     Observable<Object> startGameRequest();
   }
