+++ android/tutorials/tutorial3/src/test/java/com/uber/rib/root/logged_in/off_game/OffGameInteractorTest.java
@@ -3,7 +3,6 @@ package com.uber.rib.root.logged_in.off_game;
 import com.uber.rib.core.RibTestBasePlaceholder;
 import com.uber.rib.core.InteractorHelper;
 
-import com.uber.rib.root.logged_in.ScoreStream;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -11,15 +10,9 @@ import org.mockito.MockitoAnnotations;
 
 public class OffGameInteractorTest extends RibTestBasePlaceholder {
 
-    private final String playerOne = "playerOne";
-    private final String playerTwo = "playerTwo";
-
     @Mock OffGameInteractor.Listener listener;
     @Mock OffGameInteractor.OffGamePresenter presenter;
     @Mock OffGameRouter router;
-    @Mock ScoreStream scoreStream;
-
-
 
     private OffGameInteractor interactor;
 
@@ -27,11 +20,6 @@ public class OffGameInteractorTest extends RibTestBasePlaceholder {
     public void setup() {
         MockitoAnnotations.initMocks(this);
 
-        interactor = TestOffGameInteractor.create(
-            playerOne,
-            playerTwo,
-            listener,
-            presenter,
-            scoreStream);
+        interactor = TestOffGameInteractor.create(listener, presenter);
     }
 }
